1. Support unicode coding successfully. I have test it simply in chinese coding. 2. You can compile it with command below: gcc -O3 -Iclips -Iunicode clips/*.c unicode/*.c -o clips.out
I will complete some samples as soon as possible

+++ clips/symbol.c
@@ -647,7 +647,8 @@ globle unsigned HashInteger(
   {
    unsigned tally;
 
-   tally = (number % range);
+   /*tally = (number % range); the original commented by pandaxcl 20140802 */ 
+   tally = ((unsigned long int)number % range);
 
    return(tally);
   }
